+++ src/@constants/index.ts
@@ -7,10 +7,11 @@ export enum SVG_ELEMENTS {
     rect = 'rect'
 }
 
-export const DECIMALS = 6;
-
 export enum SVG_PROPERTIES {
     viewBox = 'viewBox'
 }
 
+export const DECIMALS = 6;
 export const SQRT3: number = +((Math.sqrt(3) / 2).toFixed(DECIMALS));
+export const DEFAULT_WIDTH = 640;
+export const DEFAULT_HEIGHT = 480;