+++ src/@demo/demo.js
@@ -5,6 +5,7 @@ import './styles.scss';
 import { Isometric } from '../';
 import demo1 from './demo1';
 import demo2 from './demo2';
+import demo3 from './demo3';
 
 const functioToString = (fn) => {
     const article = document.createElement('article');
@@ -18,7 +19,8 @@ const functioToString = (fn) => {
 
 const demos = new Map([
     ['demo1', demo1],
-    ['demo2', demo2]
+    ['demo2', demo2],
+    ['demo3', demo3]
 ]);
 
 document.addEventListener('DOMContentLoaded', function () {