+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.1.0 - 1.1.1] - 2019-12-20
+
+- Refactored how to share the canvas properties among shapes using a global class
+- Removed some internal public methods from the IsometricPath class
+- Created a new method in the Isometric path to draw using string commands
+- Created a new demo using the new method
+- New logo
+
 ## [1.0.0 - 1.0.1] - 2019-12-17
 
-- Created the isometric-js library
+- Created the isometric library